requirements.txt
python-telegram-bot
python-dotenv
requests
yt-dlp
gallery-dl
